+++ src/main/java/com/cc/blog/service/impl/UserServiceImpl.java
@@ -7,6 +7,7 @@ import com.cc.blog.util.Tools;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Service
@@ -16,7 +17,7 @@ public class UserServiceImpl implements UserService {
     private UserDao userDao;
 
     @Autowired
-    private UserServiceImpl userService;
+    private UserService userService;
 
     /**
      * 获取所有用户
@@ -24,8 +25,11 @@ public class UserServiceImpl implements UserService {
      * @return
      */
 
-    public List<User> getUserAll() {
-        return userDao.getUserAll();
+    public List<User> getUserAll(HttpServletRequest request) {
+        if(Tools.usernameSessionIsAdminValidate(request)){
+            return userDao.getUserAll();
+        }
+        return null;
     }
 
     /**
@@ -35,8 +39,11 @@ public class UserServiceImpl implements UserService {
      * @return
      */
 
-    public User getUserById(int id) {
-        return userDao.getUserById(id);
+    public User getUserById(int id,HttpServletRequest request) {
+        if(Tools.usernameSessionIsAdminValidate(request)){
+            return userDao.getUserById(id);
+        }
+        return null;
     }
 
     /**
@@ -60,8 +67,12 @@ public class UserServiceImpl implements UserService {
      * @return
      */
 
-    public Integer getUserPrivateId(String privateId) {
-        return userDao.getUserPrivateId(privateId);
+    public Integer getUserPrivateId(String privateId,HttpServletRequest request) {
+        if(Tools.usernameSessionIsAdminValidate(request)){
+            return userDao.getUserPrivateId(privateId);
+        }
+        return null;
+
     }
 
     /**
@@ -71,8 +82,12 @@ public class UserServiceImpl implements UserService {
      * @return
      */
 
-    public User getUserByUsername(String username) {
-        return userDao.getUserByUsername(username);
+    public User getUserByUsername(String username,HttpServletRequest request) {
+        if(Tools.usernameSessionIsAdminValidate(request)){
+            return userDao.getUserByUsername(username);
+        }
+        return null;
+
     }
 
     /**
@@ -81,11 +96,11 @@ public class UserServiceImpl implements UserService {
      * @param user
      */
 
-    public void insertUser(User user) {
-        Integer flag = userService.getUserPrivateId(user.getUser_common_private_id());
+    public void insertUser(User user,HttpServletRequest request) {
+        Integer flag = userService.getUserPrivateId(user.getUser_common_private_id(),request);
         while (flag == 1) {
             user.setUser_common_private_id(Tools.CreateUserRandomPrivateId());
-            flag = userService.getUserPrivateId(user.getUser_common_private_id());
+            flag = userService.getUserPrivateId(user.getUser_common_private_id(),request);
         }
         userDao.insertUser(user);
     }
@@ -96,8 +111,11 @@ public class UserServiceImpl implements UserService {
      * @param id
      */
 
-    public void deleteUserById(int id) {
-        userDao.deleteUserById(id);
+    public void deleteUserById(int id,HttpServletRequest request) {
+        if(Tools.usernameSessionIsAdminValidate(request)){
+            userDao.deleteUserById(id);
+        }
+
     }
 
     /**
@@ -106,8 +124,11 @@ public class UserServiceImpl implements UserService {
      * @param user
      */
 
-    public void updateUser(User user) {
-        userDao.updateUser(user);
+    public void updateUser(User user,HttpServletRequest request) {
+        if(Tools.usernameSessionIsAdminValidate(request)){
+            userDao.updateUser(user);
+        }
+
     }
 
     /**
